HUD refactoring
Transformed out date and welcome labels to separate class (HUDLabel).
This is primary needed to have a better layout results when resizing or
using smaller displays. Also modified line calculation so it does not
overwrite label if its text is changes.

The new HUDLabel has an embedded horizontal layout, so other icons
could be added to it later on.