Moved CS to foreground thread
There is a communication issue in the background thread making the WASM
client stalled. This was quite the opposite with previous versions of Qt.
For now, we move CS into the foreground so the development could be
continued, but this issue is kept on the TODO list, and CS would be moved
into the background thread when it is possible again.