Optimalized datapack handling [normalization]
Before we used DataBlock and DataPack. The main idea was not to expose
socket connection information to higher level executors (like Slotter
and so) The constant encapsulation and extraction generated too many
unnecessary memory calls, so all features were merged into DataPack
to simplify code and make it more error-safe.