HHC-8I8OP: Added debounce and bypass code
The HHC-8I8OP device tends to generate multiple signals upon using the
inputs leading "prelling" of the relay (quick on-off phases). The device
now debounce these quick state changes on the software side.

The device is set to bypass mode, thus the change in input always
results in state changes on the respecive relays. Bypass would be
fixed as long as the core logic is not implemented.