Added hynode.log
Currently Beacon system is not yet dispatching the version and uptime
of a device. For the POC, where there are 5+ devices online to create
the network, using file based start logging enables to see whether
a new version rollout and automatic startup is successful or not.

This "feature" would be obsoleted by beacon later on.