HHC N8I8OP: command queue implementation
According to the tests, the device does not seem to have an input queue.
Thus, if multiple command sent to the device before it processed them, the
device does nothing.

Thus a minimal queue is implemented which stores all the commands
sent to the device and send the next one only when the acknowledge
message arrived from the device.