Entity implementation approach [2]
The original approach was to create entity classes for all kind
of entities. This would enforce some naming ans methods, but
on the other side it would create a lot of constraints while loosing
some degree of control over the entity (control is more shifted from
HFS to the device/UI implementation). To make it simple, when
a device asks for an entity creation, all of the properties and so are
created and handled by the HFS itself, and a very minimal HFS
interface is provided.

This way, a lot of factory and class codes could be spared while
eliminating the risk of premature entity deletion could not be
followed up by the system.

There are other types of issues would arose, but they are handled
in specific ways.